+++ packages/widgetbook/lib/src/navigation/navigation.dart
@@ -1,4 +1,3 @@
-export 'directories/directories.dart';
 export 'icons/icons.dart';
-export 'models/models.dart';
+export 'nodes/nodes.dart';
 export 'widgets/widgets.dart';