+++ packages/widgetbook/lib/src/navigation/widgets/navigation_panel.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-import '../models/models.dart';
-import 'navigation_tree.dart';
+import '../nodes/nodes.dart';
+import 'navigation_tree_node.dart';
 import 'search_field.dart';
 
 class NavigationPanel extends StatefulWidget {
@@ -9,52 +9,83 @@ class NavigationPanel extends StatefulWidget {
     super.key,
     this.initialPath,
     this.onNodeSelected,
-    required this.directories,
+    required this.root,
   });
 
   final String? initialPath;
-  final NodeSelectedCallback? onNodeSelected;
-  final List<MultiChildNavigationNodeData> directories;
+  final ValueChanged<WidgetbookNode>? onNodeSelected;
+  final WidgetbookNode root;
 
   @override
   State<NavigationPanel> createState() => _NavigationPanelState();
 }
 
 class _NavigationPanelState extends State<NavigationPanel> {
+  late WidgetbookNode filteredRoot;
+  WidgetbookNode? selectedNode;
   String searchQuery = '';
+
+  bool filterNode(WidgetbookNode node) {
+    final regex = RegExp(
+      searchQuery,
+      caseSensitive: false,
+    );
+
+    return node.name.contains(regex);
+  }
+
+  @override
+  void initState() {
+    super.initState();
+
+    filteredRoot = widget.root;
+    selectedNode = widget.initialPath != null
+        ? widget.root.find((child) => child.path == widget.initialPath)
+        : null;
+  }
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      constraints: const BoxConstraints(minWidth: 50, maxWidth: 300),
-      child: Card(
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            const SizedBox(height: 16),
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 12),
-              child: SearchField(
-                searchValue: searchQuery,
-                onSearchChanged: (value) {
-                  setState(() => searchQuery = value);
-                },
-                onSearchCancelled: () {
-                  setState(() => searchQuery = '');
-                },
-              ),
+    return Card(
+      child: Column(
+        children: [
+          Padding(
+            padding: const EdgeInsets.all(16),
+            child: SearchField(
+              value: searchQuery,
+              onChanged: (value) {
+                setState(() {
+                  searchQuery = value;
+                  filteredRoot = widget.root.filter(filterNode) ?? widget.root;
+                });
+              },
+              onCleared: () {
+                setState(() {
+                  searchQuery = '';
+                  filteredRoot = widget.root;
+                });
+              },
             ),
-            const SizedBox(height: 16),
+          ),
+          if (filteredRoot.children != null)
             Expanded(
-              child: NavigationTree(
-                initialPath: widget.initialPath,
-                onNodeSelected: widget.onNodeSelected,
-                directories: widget.directories,
-                searchQuery: searchQuery,
+              child: ListView.builder(
+                padding: const EdgeInsets.symmetric(
+                  horizontal: 16,
+                ),
+                itemCount: filteredRoot.children!.length,
+                itemBuilder: (context, index) => NavigationTreeNode(
+                  node: filteredRoot.children![index],
+                  selectedNode: selectedNode,
+                  onNodeSelected: (node) {
+                    if (!node.isLeaf || node.path == selectedNode?.path) return;
+                    setState(() => selectedNode = node);
+                    widget.onNodeSelected?.call(node);
+                  },
+                ),
               ),
             ),
-          ],
-        ),
+        ],
       ),
     );
   }