+++ packages/widgetbook/lib/src/navigation/widgets/search_field.dart
@@ -3,37 +3,36 @@ import 'package:flutter/material.dart';
 class SearchField extends StatefulWidget {
   const SearchField({
     super.key,
-    this.onSearchPressed,
-    this.onSearchChanged,
-    this.onSearchCancelled,
-    this.searchValue = '',
+    this.onChanged,
+    this.onCleared,
+    this.value = '',
   });
 
-  final VoidCallback? onSearchPressed;
-  final ValueChanged<String>? onSearchChanged;
-  final VoidCallback? onSearchCancelled;
-  final String searchValue;
+  final ValueChanged<String>? onChanged;
+  final VoidCallback? onCleared;
+  final String value;
 
   @override
   State<SearchField> createState() => _SearchFieldState();
 }
 
 class _SearchFieldState extends State<SearchField> {
-  late final TextEditingController textEditingController;
+  late final TextEditingController controller;
   final FocusNode focusNode = FocusNode();
 
   @override
   void initState() {
-    textEditingController = TextEditingController(text: widget.searchValue);
     super.initState();
+
+    controller = TextEditingController(text: widget.value);
   }
 
   @override
   Widget build(BuildContext context) {
     return TextFormField(
-      controller: textEditingController,
+      controller: controller,
       focusNode: focusNode,
-      onChanged: widget.onSearchChanged,
+      onChanged: widget.onChanged,
       decoration: InputDecoration(
         hintText: 'Search',
         border: OutlineInputBorder(
@@ -52,22 +51,21 @@ class _SearchFieldState extends State<SearchField> {
         focusColor: Colors.white,
         iconColor: Colors.white,
         contentPadding: const EdgeInsets.symmetric(vertical: 5),
-        prefixIcon: Padding(
-          padding: const EdgeInsets.only(left: 5, right: 10),
-          child: IconButton(
-            onPressed: widget.onSearchPressed,
-            hoverColor: Colors.white.withOpacity(0.2),
-            icon: const Icon(Icons.search),
-          ),
+        prefixIcon: const Padding(
+          padding: EdgeInsets.all(12),
+          child: Icon(Icons.search),
         ),
-        suffixIcon: widget.searchValue.isNotEmpty
+        suffixIcon: widget.value.isNotEmpty
             ? Padding(
-                padding: const EdgeInsets.only(right: 5, left: 10),
+                padding: const EdgeInsets.only(
+                  right: 6,
+                  left: 12,
+                ),
                 child: IconButton(
                   onPressed: () {
-                    textEditingController.clear();
+                    controller.clear();
                     focusNode.unfocus();
-                    widget.onSearchCancelled?.call();
+                    widget.onCleared?.call();
                   },
                   hoverColor: Colors.white.withOpacity(0.2),
                   icon: const Icon(Icons.close),