+++ packages/widgetbook/lib/src/widgetbook.dart
@@ -38,7 +38,7 @@ class Widgetbook extends StatefulWidget {
   /// Both [WidgetbookCategory] and [WidgetbookFolder] can contain sub folders
   /// and [WidgetbookComponent] elements. However, [WidgetbookComponent] can
   /// only contain [WidgetbookUseCase]s.
-  final List<MultiChildNavigationNodeData> directories;
+  final List<WidgetbookNode> directories;
 
   /// A wrapper builder method for all [WidgetbookUseCase]s.
   final AppBuilder appBuilder;
@@ -55,7 +55,7 @@ class Widgetbook extends StatefulWidget {
   static Widgetbook cupertino({
     Key? key,
     String initialRoute = '/',
-    required List<MultiChildNavigationNodeData> directories,
+    required List<WidgetbookNode> directories,
     AppBuilder appBuilder = cupertinoAppBuilder,
     List<WidgetbookAddon>? addons,
     List<WidgetbookIntegration>? integrations,
@@ -74,7 +74,7 @@ class Widgetbook extends StatefulWidget {
   static Widgetbook material({
     Key? key,
     String initialRoute = '/',
-    required List<MultiChildNavigationNodeData> directories,
+    required List<WidgetbookNode> directories,
     AppBuilder appBuilder = materialAppBuilder,
     List<WidgetbookAddon>? addons,
     List<WidgetbookIntegration>? integrations,
@@ -105,7 +105,9 @@ class _WidgetbookState extends State<Widgetbook> {
       appBuilder: widget.appBuilder,
       addons: widget.addons,
       integrations: widget.integrations,
-      directories: widget.directories,
+      root: WidgetbookRoot(
+        children: widget.directories,
+      ),
     );
 
     router = AppRouter(